Request location permission when user selects auto for day/night on 23+ - closes #101
